solved day 8 part 1

+++ Model/Segment.cs
@@ -2,11 +2,11 @@ using Common;
 
 namespace Model;
 
-public struct Segment(int indexA, int indexB, long value)
+public struct Segment(int indexA, int indexB, double value)
 {
     public int A { get; set; } = indexA;
     public int B { get; set; } = indexB;
-    public long Value { get; set; } = value;
+    public double Value { get; set; } = value;
 
     public override string ToString()
     {
